refactor: move task completion into periods store and restore isInitialized state
